collector/requirements.collector.txt
requests
lxml
pandas
orjson
sqlalchemy>=2.0
psycopg[binary]>=3.1,<3.2
PyYAML
